+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = "2024, Stephen Whitlock"
 author = "Stephen Whitlock"
 
 # The full version, including alpha/beta/rc tags
-release = "0.52.0"
+release = "0.53.0"
 
 
 # -- General configuration ---------------------------------------------------